- Add option to build pipoe

+++ server/build_env.sh
@@ -6,5 +6,7 @@ source server_runtime/bin/activate
 pip3 install coloredlogs pyserial python-socketio eventlet
 pip3 freeze > requirements.txt
 
-rm -rf /work/yoctoland/meta-darksky/recipes-darksky/darksky-server-python/*
-pipoe --python python3 -r requirements.txt  -o /work/yoctoland/meta-darksky/recipes-darksky/darksky-server-python
+if [ $1 == "generate" ]; then
+    rm -rf /work/yoctoland/meta-darksky/recipes-darksky/darksky-server-python/*
+    pipoe --python python3 -r requirements.txt  -o /work/yoctoland/meta-darksky/recipes-darksky/darksky-server-python
+fi
